requirements.txt
flask-restful==0.3.10
flask==3.1.3
pyyaml==6.0.3
peewee==4.0.2
pymysql==1.1.2
pytest==9.0.2
pytest-html==4.2.0